% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PELO.R
\name{game_outcome_elo}
\alias{game_outcome_elo}
\title{Simulate an Elo game outcome (internal)}
\usage{
game_outcome_elo(p1_r, p2_r)
}
\arguments{
\item{p1_r}{Numeric. Rating of player 1.}

\item{p2_r}{Numeric. Rating of player 2.}
}
\value{
Integer. 1 if player 1 wins, 0 if player 2 wins.
}
\description{
Simulate the outcome of a match between two players using Elo ratings.
This function is intended for internal use only.
}
\keyword{internal}
