% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PELO.R
\name{game_outcome_pelo}
\alias{game_outcome_pelo}
\title{Simulate a PELO game outcome (internal)}
\usage{
game_outcome_pelo(p1_r, p2_r, p1_adv_p2)
}
\arguments{
\item{p1_r}{Numeric. Rating of player 1.}

\item{p2_r}{Numeric. Rating of player 2.}

\item{p1_adv_p2}{Numeric. Pairwise advantage of player 1 over player 2.}
}
\value{
Integer. 1 if player 1 wins, 0 if player 2 wins.
}
\description{
Simulate the outcome of a match between two players using the PELO model.
This function is intended for internal use only.
}
\keyword{internal}
