% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PELO.R
\name{get_optimized_para}
\alias{get_optimized_para}
\title{Obtain the optimized parameters under Elo/P-Elo}
\usage{
get_optimized_para(initial_guess, elo_type, data, elo_rat_KK2)
}
\arguments{
\item{initial_guess}{Initial parameters to be used}

\item{elo_type}{Elo/P-Elo}

\item{data}{Game history of all players}

\item{elo_rat_KK2}{Current ratings and pairwise advantages of all players}
}
\value{
The optimized parameters under Elo/P-Elo
}
\description{
Obtain the optimized parameters under Elo/P-Elo
}
