% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PELO.R
\name{initialize_ratings}
\alias{initialize_ratings}
\title{Initialize Ratings for Players}
\usage{
initialize_ratings(players, initial_rating_val)
}
\arguments{
\item{players}{Character vector of player names.}

\item{initial_rating_val}{Numeric value for the initial rating of each player.}
}
\value{
A dataframe with columns:
\describe{
\item{Player}{Player names.}
\item{Ratings}{Initial rating values for each player.}
}
}
\description{
Creates a dataframe of players with an initial rating value.
}
\examples{
initialize_ratings(c("Alice", "Bob"), 1500)

}
