% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PELO.R
\name{likelihood_ratio_test}
\alias{likelihood_ratio_test}
\title{Conduct the likelihood ratio test (LRT) to conclude strong/weak intransitivity}
\usage{
likelihood_ratio_test(initial_guess, data, elo_rat_KK2)
}
\arguments{
\item{initial_guess}{Initial parameters to be used}

\item{data}{Game history of all players}

\item{elo_rat_KK2}{Current ratings and pairwise advantages of all players}
}
\value{
The LRT statistic and other info: c("K_mle", "KK2_mle", "nll_elo", "nll_pelo", "lrt_statistic")
}
\description{
Conduct the likelihood ratio test (LRT) to conclude strong/weak intransitivity
}
