% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PELO.R
\name{p_1win2___KK2}
\alias{p_1win2___KK2}
\title{Calculate the winning probability of player 1}
\usage{
p_1win2___KK2(p1_r, p2_r, p1_pw, draw)
}
\arguments{
\item{p1_r}{Rating of player 1}

\item{p2_r}{Rating of player 2}

\item{p1_pw}{Pairwise advantage of of player 1 over player 2}

\item{draw}{The parameter of draw}
}
\value{
The winning probability of player 1
}
\description{
Calculate the winning probability of player 1
}
