% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PELO.R
\name{prediction}
\alias{prediction}
\title{Predict a single game result given players' ratings and pairwise advantage}
\usage{
prediction(p1_rating, p2_rating, p1_pairwise_p2, draw, prob = FALSE)
}
\arguments{
\item{p1_rating}{Rating of player 1}

\item{p2_rating}{Rating of player 2}

\item{p1_pairwise_p2}{Pairwise advantage of player 1 over player 2}

\item{draw}{The parameter for draw}

\item{prob}{Return the winning probability if TRUE, return game result prediction if FALSE}
}
\value{
The LRT statistic and other info: c("K_mle", "KK2_mle", "nll_elo", "nll_pelo", "lrt_statistic")
}
\description{
Predict a single game result given players' ratings and pairwise advantage
}
