% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PELO.R
\name{ratings_update_elo}
\alias{ratings_update_elo}
\title{Update Elo Ratings After a Match}
\usage{
ratings_update_elo(p1_r, p2_r, K, X)
}
\arguments{
\item{p1_r}{Numeric, current rating of player 1.}

\item{p2_r}{Numeric, current rating of player 2.}

\item{K}{Numeric, Elo K-factor.}

\item{X}{Integer, match outcome (1 if player 1 wins, 0 if player 2 wins).}
}
\value{
Numeric vector of updated ratings: c(new_p1_r, new_p2_r)
}
\description{
Updates the Elo ratings of two players based on the match outcome.
}
\examples{
ratings_update_elo(1600, 1500, K = 32, X = 1)

}
