% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PELO.R
\name{winning_prob_elo}
\alias{winning_prob_elo}
\title{Calculate Elo Winning Probability}
\usage{
winning_prob_elo(p1_r, p2_r)
}
\arguments{
\item{p1_r}{Numeric, rating of player 1.}

\item{p2_r}{Numeric, rating of player 2.}
}
\value{
Numeric value between 0 and 1 representing the probability of player 1 winning.
}
\description{
Computes the probability that player 1 wins against player 2 using Elo ratings.
}
\examples{
winning_prob_elo(1600, 1500)

}
