% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PELO.R
\name{winning_prob_pelo}
\alias{winning_prob_pelo}
\title{Calculate PELO Winning Probability with Pairwise Advantage}
\usage{
winning_prob_pelo(p1_r, p2_r, p1_adv_p2)
}
\arguments{
\item{p1_r}{Numeric, rating of player 1.}

\item{p2_r}{Numeric, rating of player 2.}

\item{p1_adv_p2}{Numeric, pairwise advantage of player 1 over player 2.}
}
\value{
Numeric value between 0 and 1 representing the probability of player 1 winning.
}
\description{
Computes the probability that player 1 wins against player 2 considering a pairwise advantage.
}
\examples{
winning_prob_pelo(1600, 1500, 50)

}
