% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PELO.R
\name{get_negloglik___KK2}
\alias{get_negloglik___KK2}
\title{Obtain the negative log likelihood under Elo/P-Elo}
\usage{
get_negloglik___KK2(KK2d, elo_type, data, elo_rat_KK2, output)
}
\arguments{
\item{KK2d}{Parameters to be used in this calculation}

\item{elo_type}{Elo/P-Elo}

\item{data}{Game history of all players}

\item{elo_rat_KK2}{Current ratings and pairwise advantages of all players}

\item{output}{Type of output: 'negloglik' or 'all'}
}
\value{
The negative log likelihood under Elo/P-Elo
}
\description{
Obtain the negative log likelihood under Elo/P-Elo
}
