% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PELO.R
\name{initialize_rating_table}
\alias{initialize_rating_table}
\title{Initialize a ratings table containing players' ratings and their pairwise advantages}
\usage{
initialize_rating_table(list_of_players)
}
\arguments{
\item{list_of_players}{Names of all players, user should ensure all names are syntactically safe.}

\item{initial_ratings}{Initial ratings, default value is 1500}

\item{initial_pairwise_advantages}{Initial pairwise advantages, default value is 0}
}
\value{
An initial rating table
}
\description{
Initialize a ratings table containing players' ratings and their pairwise advantages
}
