% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PELO.R
\name{rating_pairwise_update___KK2}
\alias{rating_pairwise_update___KK2}
\title{Update the ratings and pairwise advantage of player 1 and player 2}
\usage{
rating_pairwise_update___KK2(p1_r, p2_r, p1_pw, p2_pw, gameresult, K, K2, draw)
}
\arguments{
\item{p1_r}{Rating of player 1}

\item{p2_r}{Rating of player 2}

\item{p1_pw}{Pairwise advantage of player 1 over player 2}

\item{p2_pw}{Pairwise advantage of player 2 over player 1 (should be p2_pw = -p1_pw)}

\item{gameresult}{Game result (0/1)}

\item{K}{Parameter K}

\item{K2}{Parameter K2}

\item{draw}{Parameter draw}
}
\value{
The updated ratings and pairwise advantages: c(p1_r_new, p2_r_new, p1_pw_new, p2_pw_new)
}
\description{
Update the ratings and pairwise advantage of player 1 and player 2
}
